policy for discounts

+++ modual/ali/Discount/src/Http/Controllers/DiscountController.php
@@ -17,6 +17,7 @@ class DiscountController extends Controller
 
     public function index(CourseRepo $courseRepo, DiscountRepo $discountRepo)
     {
+        $this->authorize('manage', Discount::class);
         $courses = $courseRepo->getAll(Course::CONFIRMATION_STATUS_ACCEPTED);
         $discounts = $discountRepo->paginateAll();
         return view('Discount::index', compact('courses', 'discounts'));
@@ -24,7 +25,7 @@ class DiscountController extends Controller
 
     public function store(DiscountRequest $discountRequest, DiscountRepo $discountRepo)
     {
-
+        $this->authorize('manage', Discount::class);
 
         $discountRepo->Store($discountRequest->all());
 
@@ -32,11 +33,11 @@ class DiscountController extends Controller
 
         return back();
 
-
     }
 
     public function destroy(Discount $discount)
     {
+        $this->authorize('manage', Discount::class);
 
         $discount->delete();
 
@@ -47,7 +48,7 @@ class DiscountController extends Controller
 
     public function edit(Discount $discount, CourseRepo $courseRepo)
     {
-
+        $this->authorize('manage', Discount::class);
         $courses = $courseRepo->getAll(Course::CONFIRMATION_STATUS_ACCEPTED);
         return view("Discount::edit", compact("discount", "courses"));
 
@@ -57,7 +58,7 @@ class DiscountController extends Controller
     public function update(Discount $discount, DiscountRequest $discountRequest, DiscountRepo $discountRepo)
     {
 
-
+        $this->authorize('manage', Discount::class);
         $discountRepo->update($discount->id, $discountRequest->all());
         newFeedbacks();
         return redirect()->route('discounts.index');
